refactor: OnCard to separate file
