+++ OneCard.css
@@ -1,17 +1,12 @@
-.DataCard {
-  color: var(--white);
-  background-color: var(--brown-grey);
-}
-
-.DataCard > header {
+.OneCard > header {
   padding: 8px 0;
   display: flex;
   justify-content: space-around;
   align-items: center;
 }
-/* .DataCard > footer {
+.OneCard > footer {
   padding: 8px 0;
   display: flex;
   justify-content: space-around;
   align-items: center;
-} */
+}
